♻️ refactor: chat mock data를 data폴더 안으로 이동
